requirements.txt
Django==4.1.1
environs==9.5.0
more-itertools==8.14.0
phonenumbers==8.12.55
Pillow==9.2.0
python-telegram-bot==13.2
requests==2.28.1
python-dotenv~=0.21.0